Drop generalized IDocProperties adapter.
We only have one single implementation which adapts document. It is
unclear to me why we need this adapter implementation for the entry
point at the moment, also naming is a bit unclear to me. Thus i drop the
adapter in favor of a class DocPropertyCollector for now. Also we move
the doc property collection entry point to the document package as that
reflects reality at the moment.

To get doc-properties for other content we do have the
IDocPropertyProvider adapters which will be called from the
DocPropertyCollector.
