+++ opengever/dossier/tests/test_default_doc_property_adapters.py
@@ -1,4 +1,4 @@
-from opengever.dossier.interfaces import IDocProperties
+from opengever.document.docprops import DocPropertyCollector
 from opengever.dossier.interfaces import IDocPropertyProvider
 from opengever.dossier.tests import EXPECTED_DOC_PROPERTIES
 from opengever.dossier.tests import EXPECTED_DOCUMENT_PROPERTIES
@@ -8,7 +8,6 @@ from opengever.dossier.tests import EXPECTED_TASKDOC_PROPERTIES
 from opengever.dossier.tests import EXPECTED_USER_DOC_PROPERTIES
 from opengever.testing import IntegrationTestCase
 from zope.component import getAdapter
-from zope.component import getMultiAdapter
 
 
 class TestDocProperties(IntegrationTestCase):
@@ -18,27 +17,19 @@ class TestDocProperties(IntegrationTestCase):
     def test_default_doc_properties_adapter(self):
         self.login(self.regular_user)
 
-        docprops = getMultiAdapter(
-            (self.document, self.portal.REQUEST), IDocProperties)
-        all_properties = docprops.get_properties()
+        all_properties = DocPropertyCollector(self.document).get_properties()
         self.assertEqual(EXPECTED_DOC_PROPERTIES, all_properties)
 
     def test_default_doc_properties_adapter_for_taskdocument(self):
         self.login(self.regular_user)
 
-        docprops = getMultiAdapter(
-            (self.taskdocument, self.portal.REQUEST), IDocProperties)
-
-        all_properties = docprops.get_properties()
+        all_properties = DocPropertyCollector(self.document).get_properties()
         self.assertEqual(EXPECTED_TASKDOC_PROPERTIES, all_properties)
 
     def test_default_doc_properties_adapter_for_proposaldocument(self):
         self.login(self.regular_user)
 
-        docprops = getMultiAdapter(
-            (self.proposaldocument, self.portal.REQUEST), IDocProperties)
-
-        all_properties = docprops.get_properties()
+        all_properties = DocPropertyCollector(self.document).get_properties()
         self.assertEqual(EXPECTED_PROPOSALDOC_PROPERTIES, all_properties)
 
     def test_default_document_doc_properties_provider(self):
